exclude validation logic to utils
